+++ Toggly.FeatureManagement.Next/nextjs-toggly-client/src/context.tsx
@@ -281,7 +281,10 @@ function TogglyProviderOwner({
   )
 
   const getVariantValue = useCallback(
-    (featureKey: string): unknown | null => client.getVariantValue(featureKey),
+    <T = unknown>(
+      featureKey: string,
+      isT?: (v: unknown) => v is T,
+    ): T | null => client.getVariantValue(featureKey, isT),
     [client]
   )